Moved into active support namespace.
