+++ spec/lib/logstasher_spec.rb
@@ -107,8 +107,8 @@ describe LogStasher do
       expect(LogStasher).to receive(:require).with('logstasher/rails_ext/action_controller/metal/instrumentation')
       expect(LogStasher).to receive(:require).with('logstash-event')
       expect(LogStasher).to receive(:suppress_app_logs).with(app)
-      expect(LogStasher::RequestLogSubscriber).to receive(:attach_to).with(:action_controller)
-      expect(LogStasher::MailerLogSubscriber).to receive(:attach_to).with(:action_mailer)
+      expect(LogStasher::ActiveSupport::LogSubscriber).to receive(:attach_to).with(:action_controller)
+      expect(LogStasher::ActiveSupport::MailerLogSubscriber).to receive(:attach_to).with(:action_mailer)
       expect(logger).to receive(:level=).with('warn')
       LogStasher.setup(app)
       expect(LogStasher.source).to eq (logstasher_source || 'unknown')
